# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users":{"get":{"tags":["user-controller"],"summary":"전체 User 목록 조회","operationId":"findAll","responses":{"200":{"description":"User 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["user-controller"],"summary":"User 등록","operationId":"create","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"userCreateRequest":{"$ref":"#/components/schemas/UserCreateRequest"},"profile":{"type":"string","format":"binary"}},"required":["userCreateRequest"]}}}},"responses":{"201":{"description":"User가 성공적으로 생성됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"같은 email 또는 username를 사용하는 User가 이미 존재함","content":{"*/*":{"example":"User with email {email} already exists"}}}}}},"/api/readStatuses":{"get":{"tags":["read-status-controller"],"summary":"User의 Message 읽음 상태 목록 조회","operationId":"findAllByUserId","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Message 읽음 상태 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReadStatusDto"}}}}}}},"post":{"tags":["read-status-controller"],"summary":"Message 읽음 상태 생성","operationId":"readStatusCreate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadStatusCreateRequest"}}},"required":true},"responses":{"201":{"description":"Message 읽음 상태가 성공적으로 생성됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReadStatusDto"}}}},"404":{"description":"Channel 또는 User를 찾을 수 없음","content":{"*/*":{"example":"Channel | User with id {channelId | userId} not found"}}},"400":{"description":"이미 읽음 상태가 존재함","content":{"*/*":{"example":"ReadStatus with userId {userId} and channelId {channelId} already exists"}}}}}},"/api/messages":{"get":{"tags":["message-controller"],"summary":"Channel의 Message 목록 조회 성공","operationId":"findByChannelId","parameters":[{"name":"channelId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Message 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}}}}}}},"post":{"tags":["message-controller"],"summary":"Message 생성","operationId":"create_1","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"messageCreateRequest":{"$ref":"#/components/schemas/MessageCreateRequest"},"attachments":{"type":"array","items":{"type":"string","format":"binary"}}},"required":["messageCreateRequest"]}}}},"responses":{"201":{"description":"Message가 성공적으로 생성됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}},"404":{"description":"Channel 또는 User를 찾을 수 없음","content":{"*/*":{"example":"Channel | Author with id {channelId | authorId| not found"}}}}}},"/api/channels/public":{"post":{"tags":["channel-controller"],"summary":"Public Channel 생성","operationId":"publicChannelCreate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicChannelCreateRequest"}}},"required":true},"responses":{"201":{"description":"Public Channel이 성공적으로 생성됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}}},"/api/channels/private":{"post":{"tags":["channel-controller"],"summary":"Private Channel 생성","operationId":"privateChannelCreate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateChannelCreateRequest"}}},"required":true},"responses":{"201":{"description":"Private Channel이 성공적으로 생성됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"summary":"로그인","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"로그인 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"404":{"description":"사용자를 찾을 수 없음","content":{"*/*":{"example":"User with username {username} not found"}}},"400":{"description":"비밀번호가 일치하지 않음","content":{"*/*":{"example":"Wrong password"}}}}}},"/api/users/{userId}":{"delete":{"tags":["user-controller"],"summary":"User 삭제","operationId":"delete","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"User가 성공적으로 삭제됨"},"404":{"description":"User를 찾을 수 없음"}}},"patch":{"tags":["user-controller"],"summary":"User 정보 수정","operationId":"update","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"userUpdateRequest":{"$ref":"#/components/schemas/UserUpdateRequest"},"profile":{"type":"string","format":"binary"}},"required":["userUpdateRequest"]}}}},"responses":{"200":{"description":"User 정보가 성공적으로 수행됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"User를 찾을 수 없음","content":{"*/*":{"example":"user with email {email} already exists"}}}}}},"/api/users/{userId}/userStatus":{"patch":{"tags":["user-controller"],"summary":"User 온라인 상태 업데이트","operationId":"updateStatusUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"User 온라인 상태가 성공적으로 업데이트됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"해당 User의 UserStatus를 찾을 수 없음","content":{"*/*":{"example":"UserStatus with userId {userId} not found"}}}}}},"/api/readStatuses/{readStatusId}":{"patch":{"tags":["read-status-controller"],"summary":"Message 읽음 상태 수정","operationId":"Update","parameters":[{"name":"readStatusId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Message 읽음 상태가 성공적으로 수정됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReadStatusDto"}}}},"404":{"description":"Message 읽음 상태를 찾을 수 없음","content":{"*/*":{"example":"ReadStatus with id {readStatusId} not found"}}}}}},"/api/messages/{messageId}":{"delete":{"tags":["message-controller"],"summary":"Message 삭제","operationId":"delete_1","parameters":[{"name":"messageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Message가 성공적으로 삭제됨"},"404":{"description":"Message를 찾을 수 없음","content":{"*/*":{"example":"Message with id {messageId} not found"}}}}},"patch":{"tags":["message-controller"],"summary":"Message 내용 수정","operationId":"update_1","parameters":[{"name":"messageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Message가 성공적으로 수정됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDto"}}}},"404":{"description":"Message를 찾을 수 없음","content":{"*/*":{"example":"Message with id {messageId} not found"}}}}}},"/api/channels/{channelId}":{"delete":{"tags":["channel-controller"],"summary":"Channel 삭제","operationId":"channelDelete","parameters":[{"name":"channelId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Channel이 성공적으로 삭제됨"},"404":{"description":"Channel을 찾을 수 없음","content":{"*/*":{"example":"Channel with id {channelId} not found"}}}}},"patch":{"tags":["channel-controller"],"summary":"Channel 정보 수정","operationId":"channelUpdate","parameters":[{"name":"channelId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicChannelUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Channel 정보가 성공적으로 수정됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}},"404":{"description":"Channel을 찾을 수 없음","content":{"*/*":{"example":"Channel with id {channelId} not found"}}},"400":{"description":"Private Channel은 수정할 수 없음","content":{"*/*":{"example":"Channel with id {channelId} not found"}}}}}},"/api/channels":{"get":{"tags":["channel-controller"],"summary":"User가 참여중인 Channel 목록 조회","operationId":"channelFindAllByUserId","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Channel 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChannelDto"}}}}}}}},"/api/binaryContents":{"get":{"tags":["binary-content-controller"],"summary":"여러 첨부 파일 조회","operationId":"findAllByIdIn","parameters":[{"name":"binaryContentIds","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"uuid"}}}],"responses":{"200":{"description":"첨부 파일 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BinaryContentDto"}}}}}}}},"/api/binaryContents/{binaryContentId}":{"get":{"tags":["binary-content-controller"],"summary":"첨부 파일 조회","operationId":"findById","parameters":[{"name":"binaryContentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"첨부 파일 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryContentDto"}}}},"404":{"description":"첨부 파일을 찾을 수 없음","content":{"*/*":{"example":"BinaryContent with id {binaryContentId} not fond"}}}}}}},"components":{"schemas":{"UserCreateRequest":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"username":{"type":"string"},"email":{"type":"string"},"profileId":{"type":"string","format":"uuid"},"online":{"type":"boolean"}}},"ReadStatusCreateRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"channelId":{"type":"string","format":"uuid"},"lastReadAt":{"type":"string","format":"date-time"}}},"ReadStatusDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"string","format":"uuid"},"channelId":{"type":"string","format":"uuid"},"lastReadAt":{"type":"string","format":"date-time"}}},"MessageCreateRequest":{"type":"object","properties":{"content":{"type":"string"},"channelId":{"type":"string","format":"uuid"},"authorId":{"type":"string","format":"uuid"}}},"MessageDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"content":{"type":"string"},"channelId":{"type":"string","format":"uuid"},"authorId":{"type":"string","format":"uuid"},"attachmentIds":{"type":"array","items":{"type":"string","format":"uuid"}}}},"PublicChannelCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"ChannelDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["PUBLIC","PRIVATE"]},"name":{"type":"string"},"description":{"type":"string"},"participantIds":{"type":"array","items":{"type":"string","format":"uuid"}},"lastMessageAt":{"type":"string","format":"date-time"}}},"PrivateChannelCreateRequest":{"type":"object","properties":{"participantIds":{"type":"array","items":{"type":"string","format":"uuid"}}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"UserUpdateRequest":{"type":"object","properties":{"newUsername":{"type":"string"},"newEmail":{"type":"string"},"newPassword":{"type":"string"}}},"UserStatusUpdateRequest":{"type":"object","properties":{"newLastActiveAt":{"type":"string","format":"date-time"}}},"ReadStatusUpdateRequest":{"type":"object","properties":{"newLastReadAt":{"type":"string","format":"date-time"}}},"MessageUpdateRequest":{"type":"object","properties":{"newContent":{"type":"string"}}},"PublicChannelUpdateRequest":{"type":"object","properties":{"newName":{"type":"string"},"newDescription":{"type":"string"}}},"BinaryContentDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"fileName":{"type":"string"},"size":{"type":"integer","format":"int64"},"contentType":{"type":"string"},"bytes":{"type":"string","format":"byte"}}}}}}